pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "looker-fields"
version = "0.1.0"
description = "Extract field-level metadata from any Looker instance via the API"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
authors = [
    {name = "Joon Solutions"},
]
dependencies = [
    "httpx[http2]>=0.27",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "typer>=0.12",
    "orjson>=3.10",
    "pyarrow>=15.0",
    "google-cloud-bigquery>=3.20",
    "looker-sdk>=24.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[project.scripts]
looker-fields = "looker_fields.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/looker_fields"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
